Organized the tests into one file
