+++ test/integrationTests.js
@@ -1,0 +1,9 @@
+require('./dbMock');
+
+const server = require('../server');
+
+module.exports = () => {
+    require('./wordRoutesTests')(server);
+    require('./userRoutesTests')(server);
+    require('./gameDataRoutesTests')(server);
+}
