+++ test/tests.test.js
@@ -1,0 +1,7 @@
+describe('Unit tests', () => {
+    //unit tests here
+});
+
+describe('Integration tests', () => {
+    require('./integrationTests')();
+});
